requirements.txt
aiosmtplib==5.1.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.17
decorator==5.2.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
executing==2.2.1
fastapi-mail==1.6.1
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
six==1.17.0
stack-data==0.6.3
starlette==0.52.1
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
